% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addNoise}
\alias{addNoise}
\title{Add noise to gene expression data}
\usage{
addNoise(x = NULL, pt = 0.1, type = "NB")
}
\arguments{
\item{x}{a gene expression numeric vector.}

\item{pt}{parameter to control noised level. Default: 0.1}

\item{type}{"NB", "N" or "LN". "NB" means Negative binomial model,
"N" means normal model, "LN" means Log-normal model.
Default: "NB"}
}
\value{
the sample length vector.
}
\description{
Add noise based on negtive bionomial distribution. Please see
"A benchmark for RNA-seq deconvolution analysis
under dynamic testing environments" in Genome Biology.
}
\examples{
res <- addNoise(x = seq(100))

}
