% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoiseExpr.R
\name{addNoiseExpr}
\alias{addNoiseExpr}
\title{Add noise to the simulated expression matrix}
\usage{
addNoiseExpr(
  exprFile,
  outputPath = NULL,
  prefix = NULL,
  Pt = seq(0.1, 1, 0.1),
  type = "NB"
)
}
\arguments{
\item{exprFile}{The input expression file. Must be in csv format. Each row is
a gene, each column is a sample. rownames and colnames are required. Please check
the output of the function \code{\link{exprSim}}.}

\item{outputPath}{Output path, create if not exists. Default: a new folder
based on the exprFile. For example, if the exprFile is "/data/test.txt",
the outputPath will be a new folder "/data/test".}

\item{prefix}{The prefix of the output file.
Note: the suffix will be added automatically based on the noise level.}

\item{Pt}{Parameter to control noise level. Default: seq(0.1, 1, 0.1)}

\item{type}{"NB", "N" or "LN". Noise type. "NB" means Negative binomial model,
"N" means normal model, "LN" means Log-normal model.
Default: "NB"}
}
\value{
All the information will be written in the output path, and each file
is the generated data in a specific noise level.
}
\description{
Add different level noise to the simulated expression matrix
based on the negative binomial distribution. The related model can be found
in this paper (Jin, H., Liu, Z. A benchmark for RNA-seq deconvolution analysis
 under dynamic testing environments. Genome Biol 22, 102 (2021).
 https://doi.org/10.1186/s13059-021-02290-6).
}
\examples{
res <- pseudoData(type = 2)
addNoiseExpr(exprFile = res$actual)

}
