% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_noise.R
\name{boxplot_NcrossCompare}
\alias{boxplot_NcrossCompare}
\title{Comparing boxplot for noise testing.}
\usage{
boxplot_NcrossCompare(
  actual,
  predicted,
  label = NULL,
  method,
  title = "Boxplot"
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{a list of vectors. Each vector contains all the files with
the predicted proportions in different noise level. Must be in csv file with
row and column names.
row: cell types, column: samples.
For example, list(method1 = c("m1_noise0.csv", "m1_noise1.csv", "m1_noise2.csv"),
method2 = c("m2_noise0.csv", "m2_noise1.csv", "m2_noise2.csv")).
Note: names for each method will be used for assign colors.}

\item{label}{a vector contains the label corresponding to the predicted
proportion file. Default: c("condition1", "condition2", "condition3", ...)}

\item{method}{"rmse", "mape", "mae", "pearson" or "spearman".
Note: for mape, cell types with read proportion 0 will be ignored.}

\item{title}{Boxplot title in character. Default: "Boxplot".}
}
\value{
The computed metrics and plot data.
}
\description{
Boxplot for illustrating the deconvlition performance with the
noised in silico data. this function is used for comparing multiple methods.
}
