% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplotR}
\alias{boxplotR}
\title{Boxplot for deconvolution.}
\usage{
boxplotR(actual, predicted, method = NULL)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{The predicted proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{method}{"rmse", "mape", "mae", "pearson" or "spearman".
Note: for mape, cell types with read proportion 0 will be ignored.}
}
\value{
The computed metrics and plot data.
}
\description{
Boxplot for deconvolution results with multiple samples.
}
