% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplotR2}
\alias{boxplotR2}
\title{Boxplot for noise testing.}
\usage{
boxplotR2(actual, predicted, label = NULL, method, title = "Boxplot")
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{a vector contains all the files with the predicted proportions
in different noise level. Must be in csv file with row and column names.
row: cell types, column: samples.
For example, c("noise0.csv", "noise1.csv", "noise2.csv").}

\item{label}{a vector contains the label corresponding to the predicted
proportion file. Default: base file name in parameter 'predicted'.
For example, c("noise0", "noise1", "noise2").}

\item{method}{"rmse", "mape", "mae", "pearson" or "spearman".
Note: for mape, cell types with read proportion 0 will be ignored.}

\item{title}{Boxplot title in character.}
}
\value{
The computed metrics and plot data.
}
\description{
Boxplot for illustrating the deconvlition performance with the
noised in silico data.
}
