% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_probabilities}
\alias{calculate_probabilities}
\title{calculate_probabilities}
\usage{
calculate_probabilities(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
prob_matrix
}
\description{
calculate_probabilities
}
\keyword{internal}
