% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rare.R
\name{cheatmap_RcrossCompare}
\alias{cheatmap_RcrossCompare}
\title{Plot circle heatmap for deconvolution results rare cell types.}
\usage{
cheatmap_RcrossCompare(
  actual,
  predicted,
  p_rare = c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05),
  label = NULL,
  method1,
  method2
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{a vector contains all the files with the predicted proportions
in different method. Must be in csv file with row and column names.
row: cell types, column: samples.
For example, c("method1.csv", "method2.csv", "method3.csv").}

\item{p_rare}{A vector of proportions. should be the same with function \code{\link{rareExprSim}}.
Default: c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05)}

\item{label}{a vector contains the label corresponding to the predicted
proportion file. Default: c("condition1", "condition2", "condition3", ...)}

\item{method1}{"rmse", "mape", "mae".
Note: for mape, cell types with read proportion 0 will be ignored.}

\item{method2}{"rmse", "mape", "mae".
Note: for mape, cell types with read proportion 0 will be ignored.}
}
\value{
The plot data.
}
\description{
Generate circle heatmap for different rare cell type proportions.
The dataset should be generated from the function \code{\link{rareExprSim}}.
}
\examples{
rareExprSim()
# after the simulation, just pass the predicted file
# to the parameter "actual" and "predicted"

}
