% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_matrix_names}
\alias{check_matrix_names}
\title{check_matrix_names}
\usage{
check_matrix_names(matrix1, matrix2)
}
\arguments{
\item{matrix1}{matrix}

\item{matrix2}{matrix}
}
\value{
bool
}
\description{
check_matrix_names
}
\keyword{internal}
