% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compute_ssim}
\alias{compute_ssim}
\title{compute ssim}
\usage{
compute_ssim(x, y, C1 = 0.01, C2 = 0.03)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\value{
output
}
\description{
compute ssim
}
\keyword{internal}
