% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{G}
\alias{G}
\title{Generate a gamma random number}
\usage{
G(shape, scale)
}
\arguments{
\item{shape}{shape}

\item{scale}{scale}
}
\value{
random number from gamma distribution
}
\description{
Generate a gamma random number.
}
\keyword{internal}
