% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getInput}
\alias{getInput}
\title{Get input matrix data}
\usage{
getInput(data = NULL, name = NULL)
}
\arguments{
\item{data}{Input}

\item{name}{variable name}
}
\value{
output
}
\description{
Get input data
}
\keyword{internal}
