% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_noise.R
\name{heatmap_NGradCT}
\alias{heatmap_NGradCT}
\title{Plot heatmap for different cell types.}
\usage{
heatmap_NGradCT(actual, predicted, label = NULL, method)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{a vector contains all the files with the predicted proportions
in different noise level. Must be in csv file with row and column names.
row: cell types, column: samples.
For example, c("noise0.csv", "noise1.csv", "noise2.csv").}

\item{label}{a vector contains the label corresponding to the predicted
proportion file. Default: c("condition1", "condition2", "condition3", ...)}

\item{method}{"rmse", "mape", "mae", "pearson" or "spearman".
Note: for mape, cell types with read proportion 0 will be ignored.}
}
\value{
The computed metrics and plot data.
}
\description{
Generate heatmap comparison for different cell types.
}
