% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kl_divergence}
\alias{kl_divergence}
\title{kl_divergence}
\usage{
kl_divergence(p, q)
}
\arguments{
\item{p}{matrix}

\item{q}{matrix}
}
\value{
matrix
}
\description{
kl_divergence
}
\keyword{internal}
