% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge.all}
\alias{merge.all}
\title{Merge dataframes}
\usage{
\method{merge}{all}(x, ..., by = "row.names")
}
\arguments{
\item{x}{A dataframe}

\item{...}{Other dataframes}

\item{by}{Flag used to merge, default: "row.names"}
}
\value{
A dataframe
}
\description{
Merge and align dataframes.
}
\keyword{internal}
