% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{multiClassSample}
\alias{multiClassSample}
\title{Split a dataframe into train and test set.}
\usage{
multiClassSample(
  df = NULL,
  p = NULL,
  labelCol = "Required_Cell_Type",
  outVarCol = "RUN_Accession"
)
}
\arguments{
\item{df}{A dataframe for all cell type data.}

\item{p}{Proportion of sample in train set.}

\item{labelCol}{column name to record cell types.}

\item{outVarCol}{column name to output, usually is sample name like RUN Accession Number.}
}
\value{
A list contains train and sample names.
}
\description{
Split a dataframe into train and test set based on the labelCol,
1 sample in one class at least.
}
\keyword{internal}
