% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple2.R
\name{plot_multiple2}
\alias{plot_multiple2}
\title{Plotting function for comparison multiple deconvolution method.}
\usage{
plot_multiple2(
  actual,
  predicted,
  condition = NULL,
  method = NULL,
  method2 = NULL,
  type = "sample",
  figure = "boxplot",
  errbar = NULL
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{a list of vectors. Each vector contains all the files with
the predicted proportions in different scenarios. Must be in csv file with
row and column names.
row: cell types, column: samples.
For example, list(method1 = c("m1_noise0.csv", "m1_noise1.csv", "m1_noise2.csv"),
method2 = c("m2_noise0.csv", "m2_noise1.csv", "m2_noise2.csv")).
Note: the name of each method should be specified. This name will be used for
generating the label when plotting.}

\item{condition}{a vector contains the condition labels corresponding to the predicted
proportion file.
For example, c("noise0", "noise1", "noise2").}

\item{method}{One of the c("mae", "rmse", "mape", "kendall", "pearson", "spearman").
Note: for mape, cell types with real proportion 0 will be ignored.}

\item{method2}{One of the c("mae", "rmse", "mape", "kendall", "pearson", "spearman").
Generating the second metric in cheatmap.}

\item{type}{One of the c("sample", "celltpye", "all").
Note: method2 will be disabled when figure is in c("boxplot", "barplot", "heatmap")}

\item{figure}{One of the c("boxplot", "barplot", "heatmap", "cheatmap")}

\item{errbar}{error bar type for barplot. One of the c("SD", "SE"), default: SE.
SD: Standard Deviation
SE: Standard Error}
}
\value{
The computed metrics and plot data.
}
\description{
Generating plots for multiple deconvolution method. This method
is designed for comparing the results from different methods under different
scenario. For example, comparing the deconvolution effect of different methods
from various noise level data. Of course, there can be many samples in a
certain scenario. But for comparing the results from data with a certain noise
level, please use \code{\link{plot_multiple}}.
Note: Function \code{\link{plot_multiple}} can reveal the deconvolution effect
for celltypes as well as samples directly. Function \code{\link{plot_multiple2}} reveals
the overall deconvolution results for different scenarios. The celltype or sample
specific effect can not be illustrated. Users should choose the appropriate
function. Of course, users can adopt the results from each function to
perform customized analysis.
}
