% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rare.R
\name{plot_rare}
\alias{plot_rare}
\title{Scatter plot for rare component.}
\usage{
plot_rare(
  actual,
  predicted,
  p_rare = c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05),
  method = NULL,
  method2 = NULL,
  type = NULL,
  celltype = TRUE,
  errbar = "SE",
  figure = NULL
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{The predicted proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.
Note: For parameter 'figure' is scatterplot, 'predicted' must be one matrix or filename.
This means scatterplot is designed for generating plot for one method.
For example, predicted = 'method1_predicted.csv'
Note: For parameter 'figure' is in c("heatmap", "cheatmap"), 'predicted' must be a vector of filename.
This means heatmap is designed for generating plot for multiple methods.
For example, predicted = c('method1_predicted.csv', 'method2_predicted.csv', 'method3_predicted.csv')}

\item{p_rare}{A vector of proportions. should be the same with function \code{\link{rareExprSim}}.
Default: c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05)}

\item{method}{One of the c("mae", "rmse", "mape", "kendall", "pearson", "spearman").
For parameter 'figure' is scatterplot, method must be one of c("kendall", "pearson", "spearman")
Note: kendall, pearson and spearman is not supported by heatmap and cheatmap.}

\item{method2}{One of the c("mae", "rmse", "mape", "kendall", "pearson", "spearman").
Generating the second metric in cheatmap.
Note: method2 will be disabled when figure is in c("scatterplot", "heatmap")
Note: kendall, pearson and spearman is not supported by heatmap and cheatmap.}

\item{type}{Not used. All metrics will be computed for each rare component,
regardless of sample and cell type.}

\item{celltype}{TRUE or FALSE. Assign different type and color for different
cell types. Default: TRUE.
Note: This parameter only supported by the scatterplot.}

\item{errbar}{standard error.}

\item{figure}{One of the c("scatterplot", "heatmap", "cheatmap", "barplot").
For different type of figure, users should pass the correct parameters.}

\item{label}{a vector contains the label corresponding to the predicted
proportion file from different methods. Default: base file name in parameter 'predicted'.
For example, c("method1", "method2", "method3").
Note: this parameter is only supported by the heatmap and cheatmap functions.}
}
\value{
The plot data.
}
\description{
Generate scatter plot for rare component.
Note: only the rare proportion will be considered.
The 'scatterplot' function can be used to estimate the deconvolution power for each cell type.
The 'heatmap' and 'cheatmap' functions can be used to compare the deconvolution power for each method.
}
\examples{
rareExprSim()
# after the simulation, just pass the predicted file
# to the parameter "actual" and "predicted"

}
