% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single.R
\name{plot_single}
\alias{plot_single}
\title{Plotting function for a single deconvolution method.}
\usage{
plot_single(
  actual,
  predicted,
  method = NULL,
  type = "sample",
  figure = "boxplot",
  errbar = "SE"
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types.
Can be a matrix or csv file with row and column names.
row: cell types, column: samples.}

\item{predicted}{The predicted proportion of cell types.
Can be a matrix or csv file with row and column names.
row: cell types, column: samples.}

\item{method}{One of the c("mae", "rmse", "mape", "kendall", "pearson", "spearman").
Note: for mape, cell types with real proportion 0 will be ignored.
Note: for scatter plot, method must in c("kendall", "pearson", "spearman").}

\item{type}{One of the c("sample", "celltpye", "all").
For "sample", generate metric for each sample. Scatter plot will assign different shape and color for each sample.
For "celltype", generate metric for each cell type. Scatter plot will assign different shape and color for each celltype.
For "all", generate metric for all data, which means flattening all data into an vector. Scatter plot will remove shape and color.
Note: boxplot is not supported for type="all".}

\item{figure}{One of the c("boxplot", "barplot", "scatterplot")
Note: boxplot is not supported for type="all".
Note: for scatter plot, method must in c("kendall", "pearson", "spearman").}

\item{errbar}{error bar type for barplot. One of the c("SD", "SE"), default: SE.
SD: Standard Deviation
SE: Standard Error}
}
\value{
The computed metrics as well as the plot data.
}
\description{
Gnerating plots for a specific deconvolution method.
}
\examples{
res <- pseudoData(type = 1)
res <- plot_single(actual = res$actual, predicted = res$predicted, method = "mape", type = "sample", figure = "boxplot")

}
