% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoData.R
\name{pseudoData}
\alias{pseudoData}
\title{Generate pseudo data for plotting}
\usage{
pseudoData(type = 1, outputPath = "test_file")
}
\arguments{
\item{type}{1, 2 or 3 for different usage}

\item{outputPath}{path to save tmp file}
}
\value{
The pseudo data
}
\description{
Generate pseudo data (prediction values) for plotting
}
\examples{
res <- pseudoData(type = 1)

}
