% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoExpr.R
\name{pseudoExpr}
\alias{pseudoExpr}
\title{Generate pseudo mixture expression matrix}
\usage{
pseudoExpr(n_sample = 50, n_gene = 1000, n_ct = 10)
}
\arguments{
\item{n_sample}{Sample number to be generated, default: 50}

\item{n_gene}{Gene number to be generated, default: 1000}

\item{n_ct}{Cell type number to be generated, default: 10}
}
\value{
a list, prop means the pseudo proportion, ref means the pseudo
external reference, mix is the output of ref %*% prop
}
\description{
Generate pseudo mixture expression matrix based on uniform
distribution, without any noise.
}
\examples{
res <- pseudoExpr()

}
