% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{refGenerator}
\alias{refGenerator}
\title{Generate reference matrix.}
\usage{
refGenerator(df = NULL, attr = NULL, cellTypes = NULL)
}
\arguments{
\item{df}{A dataframe for cell type expression data.}

\item{attr}{Attribute information.}

\item{cellTypes}{Cell types information.}
}
\value{
A reference list contains expression and expression variance matrix.
}
\description{
Generate reference matrix for deconvolution. This function computes
the expression and expression variance matrix for different cell types.
}
\keyword{internal}
