% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regMetrics.R
\name{regMetrics}
\alias{regMetrics}
\title{Compute regression Metrics for matrix results.}
\usage{
regMetrics(actual, predicted, method = NULL, type = NULL)
}
\arguments{
\item{actual}{The groundcloth proportion of cell types in matrix.
row: cell types, column: samples.}

\item{predicted}{The predicted proportion of cell types in matrix.
row: cell types, column: samples.}

\item{method}{One of c("rmse", "mape", "mae", "smape", "kendall", "pearson", "spearman"),}

\item{type}{One of c("sample.", "celltype", "all").
For "sample.", generate metric for each sample.
For "celltype", generate metric for each cell type.
For "all", generate metric for all data, which means flattening all data into an vector.
Note: for mape, cell types with read proportion 0 will be ignored.}
}
\value{
a vector with the value for each sample.
}
\description{
Compute several regression Metrics for cell type deconcolution.
}
\examples{
res <- regMetrics(actual = matrix(data = seq(9), nrow = 3),
                  predicted = matrix(data = seq(9), nrow = 3),
                  method = "rmse")
}
