% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorder_df}
\alias{reorder_df}
\title{Re-order dataframe}
\usage{
reorder_df(df1 = NULL, df2 = NULL)
}
\arguments{
\item{df1}{A dataframe1.}

\item{df2}{Attribute dataframe2.}
}
\value{
Re-order dataframe2.
}
\description{
Re-order dataframe2 by dataframe1.
}
\keyword{internal}
