% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rare.R
\name{scatter_R}
\alias{scatter_R}
\title{Scatter plot for rare component.}
\usage{
scatter_R(
  actual,
  predicted,
  p_rare = c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05),
  celltype = TRUE
)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{The predicted proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{p_rare}{A vector of proportions. should be the same with function \code{\link{rareExprSim}}.
Default: c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05)}

\item{celltype}{TRUE or FALSE. Assign different type and color for different
cell types. Default: TRUE.}
}
\value{
The plot data.
}
\description{
Generate scatter plot for rare component.
}
\examples{
rareExprSim()
# after the simulation, just pass the predicted file
# to the parameter "actual" and "predicted"

}
