% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot for proportion prediction.}
\usage{
scatterPlot(actual, predicted, method, celltype = TRUE)
}
\arguments{
\item{actual}{The groundtruth proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{predicted}{The predicted proportion of cell types in matrix or csv file.
row: cell types, column: samples. Also can be a csv matrix with row and
column names.}

\item{method}{"pearson" (default), "kendall", or "spearman".}

\item{celltype}{TRUE or FALSE. Assign different type and color for different
cell types.}
}
\value{
The plot data.
}
\description{
Generate scatter plot comparison for different cell types.
}
