% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{v_norm}
\alias{v_norm}
\title{Normalize a vector}
\usage{
v_norm(x = NULL, scale = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{scale}{a number to scaled}
}
\value{
Normalized numeric vector
}
\description{
Normalize a vector
}
\examples{
res <- v_norm(seq(100))

}
