% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoise.R, R/utilities.R
\name{addNoise}
\alias{addNoise}
\title{Add noise to the simulated expression matrix}
\usage{
addNoise(x = NULL, pt = 0.1)

addNoise(x = NULL, pt = 0.1)
}
\arguments{
\item{x}{a gene expression numeric vector.}

\item{pt}{parameter to control noised level. Default: 0.1}

\item{exprFile}{The input expression file. Must be in csv format. Each row is
a gene, each column is a sample. rownames and colnames are required. Please check
the output of the function \code{\link{exprSim}}.}

\item{outputPath}{Output path, create if not exists. Default: a new folder
based on the exprFile. For example, if the exprFile is "/data/test.txt",
the outputPath will be a new folder "/data/test".}

\item{prefix}{The prefix of the output file.
Note: the suffix will be added automatically based on the noise level.}

\item{Pt}{Parameter to control noise level. Default: seq(0.1, 1, 0.1)}
}
\value{
All the information will be written in the output path. Each file
is the generated data in a specific noise level.

the sample length vector.
}
\description{
Add different level noise to the simulated expression matrix
based on the negative binomial distribution. The related model can be found
in this paper (Jin, H., Liu, Z. A benchmark for RNA-seq deconvolution analysis
 under dynamic testing environments. Genome Biol 22, 102 (2021).
 https://doi.org/10.1186/s13059-021-02290-6).

Add noise based on negtive bionomial distribution. Please see
"A benchmark for RNA-seq deconvolution analysis
under dynamic testing environments" in Genome Biology.
}
