% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarescExprSim.R
\name{rarescExprSim}
\alias{rarescExprSim}
\title{Generate in silico mixture expression matrix based on scRNA-seq with rare component.}
\usage{
rarescExprSim(
  cell_number = 2000,
  p = 2/3,
  transform = "TPM",
  outputPath = NULL,
  bulk_name = "scPBMC_gene_expr.csv",
  ref_bulk_name = "scPBMC_ref.csv",
  ref_cell_number = 1000,
  ref_sc_name = "scPBMC_ref_sc.csv",
  ref_sc_label = "scPBMC_ref_sc_label.csv",
  prop_name = "scPBMC_prop.csv",
  type = "human_PBMC",
  p_rare = c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05)
)
}
\arguments{
\item{p}{Proportion of sample in train set, default: 2 / 3.}

\item{transform}{"TPM", "CPM" or "NO". Transform the data into TPM, CPM or in raw counts.
The suffix will be added to the file name.}

\item{outputPath}{output file save path.}

\item{bulk_name}{Pseudo bulk output file name.}

\item{ref_bulk_name}{reference output file name in csv, for the deconvolution method based on bulk data.}

\item{ref_cell_number}{number of cells of each cell type for generating reference, default: 1000.}

\item{ref_sc_name}{File name for all data in train set in csv.
This data can be used for differential gene analysis.}

\item{ref_sc_label}{cell labels for ref_sc_name.}

\item{prop_name}{simulated proportion file name in csv.}

\item{type}{'mouse_tissue' or 'human_PBMC'. Default: 'human_PBMC'}

\item{p_rare}{A vector of proportions.
Default: c(0.001, 0.003, 0.005, 0.008, 0.01, 0.03, 0.05)
Note: Every cell type will be treated as rare component.
For example, if 8 cell types need to be tested, this function will generate
7 * 8 = 56 samples. 7 means 7 rare proportions and 8 means 8 cell types.}
}
\value{
All the information will be written in the output path.
}
\description{
Generate in silico mixture expression matrix based on the
internal scRNA-seq database. The internal RNA-seq database is collected from
PMID:29474909. All the samples are passed the quality filter.
}
\examples{
rarescExprSim()

}
