% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rmse}
\alias{mape}
\alias{mae}
\alias{fread}
\alias{fwrite}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}}

  \item{Metrics}{\code{\link[Metrics]{mae}}, \code{\link[Metrics]{mape}}, \code{\link[Metrics]{rmse}}}
}}

