% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regMetrics.R
\name{regMetrics}
\alias{regMetrics}
\title{Compute regression Metrics for matrix results.}
\usage{
regMetrics(actual, predicted, method = NULL)
}
\arguments{
\item{actual}{The groundcloth proportion of cell types in matrix.
row: cell types, column: samples.}

\item{predicted}{The predicted proportion of cell types in matrix.
row: cell types, column: samples.}

\item{method}{"rmse", "mape", "mae", "pearson", "spearman", "pearson2", "spearman2".
Note: for mape, cell types with read proportion 0 will be ignored.
For pearson2 and spearman2, all data in matrix will be taken into
consideration, which mean only one number will be reported.}
}
\value{
a vector with the value for each sample.
}
\description{
Compute several regression Metrics for cell type deconcolution.
}
