% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scExprSim.R
\name{scExprSim}
\alias{scExprSim}
\title{Generate in silico mixture expression matrix based on scRNA-seq}
\usage{
scExprSim(
  n_sample = 50,
  p = 2/3,
  transform = "TPM",
  outputPath = NULL,
  mix_name = "scMouse_gene_expr.csv",
  ref_name = "scMouse_ref.csv",
  prop_name = "scMouse_prop.csv",
  train_name = "scMouse_ref_rawCount.csv"
)
}
\arguments{
\item{n_sample}{Sample number to be generated, default: 50.}

\item{p}{Proportion of sample in train set, default: 2 / 3.}

\item{transform}{"TPM", "CPM" or "NO". Transform the data into TPM, CPM or in counts.}

\item{outputPath}{output file save path.}

\item{mix_name}{mixture output file name.}

\item{ref_name}{reference output file name in csv.}

\item{prop_name}{simulated proportion file name in csv.}

\item{train_name}{file name for all data in train set in csv.
This data can be used for differential gene analysis.}
}
\value{
All the information will be written in the output path.
}
\description{
Generate in silico mixture expression matrix based on the
internal scRNA-seq database. The internal RNA-seq database is collected from
PMID:29474909. All the samples are passed the quality filter.
This function use the real cell type specific expression data to generate
mixture data. The cell type include c("FetalStomach", "FetalLung",
"FetalLiver", "FetalKidney", "FetalIntestine",
"FetalBrain", "Female.fetal.Gonad")
}
\examples{
res <- scExprSim()

}
